requirements.txt
# FastAPI & ASGI server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Base de datos con SQLAlchemy y PostgreSQL
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# Cargar variables de entorno
python-dotenv==1.1.0

# Validación de datos
pydantic==2.7.1
pydantic-settings==2.2.1

# Autenticación con JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Migraciones de base de datos
alembic==1.13.1

# CORS (para conectar frontend React con FastAPI)
fastapi[all]==0.111.0

# Utilidades opcionales (seguridad, hashing, fechas)
bcrypt==4.1.3
python-multipart==0.0.9
